restructured project to match convention and eliminating the need to recompile the same source code for both library and binary
